P8-6 crud and tests task

+++ src/Controller/TaskController.php
@@ -3,8 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Entity\User;
 use App\Form\TaskType;
 use App\Repository\TaskRepository;
+use App\Security\TaskVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
@@ -23,18 +25,26 @@ class TaskController extends AbstractController
     #[Route('/tasks', name: 'task_list')]
     public function listAction(): Response
     {
-        return $this->render('task/list.html.twig', ['tasks' => $this->taskRepository->findAll()]);
+        $this->denyAccessUnlessGranted(TaskVoter::TASK_LIST);
+
+        /** @var User $user */
+        $user = $this->getUser();
+        $tasks = $this->taskRepository->getTasksByUser($user);
+
+        return $this->render('task/list.html.twig', ['tasks' => $tasks]);
     }
 
     #[Route('/tasks/create', name: 'task_create')]
     public function createAction(Request $request): RedirectResponse|Response
     {
+        $this->denyAccessUnlessGranted(TaskVoter::TASK_CREATE);
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $task->setUser($this->getUser());
             $this->entityManager->persist($task);
             $this->entityManager->flush();
 
@@ -49,8 +59,9 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/edit', name: 'task_edit')]
     public function editAction(Task $task, Request $request): RedirectResponse|Response
     {
-        $form = $this->createForm(TaskType::class, $task);
+        $this->denyAccessUnlessGranted(TaskVoter::TASK_EDIT, $task);
 
+        $form = $this->createForm(TaskType::class, $task);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -70,9 +81,12 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
     public function toggleTaskAction(Task $task): RedirectResponse
     {
-        $task->toggle(!$task->isDone());
+        $this->denyAccessUnlessGranted(TaskVoter::TASK_EDIT, $task);
 
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        $task->toggle(!$task->isDone());
+        $this->entityManager->flush();
+
+        $this->addFlash('success', sprintf('Le statut de la tâche %s a bien été modifié.', $task->getTitle()));
 
         return $this->redirectToRoute('task_list');
     }
@@ -80,6 +94,8 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTaskAction(Task $task): RedirectResponse
     {
+        $this->denyAccessUnlessGranted(TaskVoter::TASK_DELETE, $task);
+
         $this->entityManager->remove($task);
         $this->entityManager->flush();
